proposed refactor
