+++ src/scopes/grants/index.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/prefer-default-export */
-import { createFiltersToScopes } from '../utils';
-import { activeBefore, activeAfter, activeWithinDates } from './startDate';
+import { map, pickBy } from 'lodash';
+import { activeBefore, activeAfter, activeWithinDates } from './activeWithin';
 import { withRegion, withoutRegion } from './region';
 import { withRecipientName, withoutRecipientName } from './recipient';
 import { withProgramSpecialist, withoutProgramSpecialist } from './programSpecialist';
@@ -28,13 +28,7 @@ export const topicToQuery = {
   stateCode: {
     ctn: (query) => withStateCode(query),
   },
-  startDate: {
-    bef: (query) => activeBefore(query),
-    aft: (query) => activeAfter(query),
-    win: (query) => activeWithinDates(query),
-    in: (query) => activeWithinDates(query),
-  },
-  endDate: {
+  activeWithin: {
     bef: (query) => activeBefore(query),
     aft: (query) => activeAfter(query),
     win: (query) => activeWithinDates(query),
@@ -46,6 +40,28 @@ export const topicToQuery = {
   },
 };
 
-export function grantsFiltersToScopes(filters) {
-  return createFiltersToScopes(filters, topicToQuery);
+export function grantsFiltersToScopes(filters, subset) {
+  const validFilters = pickBy(filters, (query, topicAndCondition) => {
+    const [topic, condition] = topicAndCondition.split('.');
+
+    if ((topic === 'startDate' || topic === 'endDate') && subset) {
+      return condition in topicToQuery.activeWithin;
+    }
+
+    if (!(topic in topicToQuery)) {
+      return false;
+    }
+
+    return condition in topicToQuery[topic];
+  });
+
+  return map(validFilters, (query, topicAndCondition) => {
+    const [topic, condition] = topicAndCondition.split('.');
+
+    if ((topic === 'startDate' || topic === 'endDate') && subset) {
+      return topicToQuery.activeWithin[condition]([query].flat());
+    }
+
+    return topicToQuery[topic][condition]([query].flat());
+  });
 }
